api/app/requirements.txt
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
click==7.1.1
Flask==1.1.1
flask-mongoengine==0.9.5
flask-restplus==0.13.0
Flask-WTF==0.14.3
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.19.1
pylint==2.4.4
pymongo==3.10.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
typed-ast==1.4.1
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
zipp==3.1.0
